refactor: extract EditMenu's toggle catalog and time-of-day bucketing
Move the declarative toggle catalog (the Toggle record + TOGGLES map) into a
dedicated EditMenuToggles holder, with each Toggle rendering itself, and lift
the TimeOfDay enum to its own file. EditMenu keeps the slot wiring, the
heterogeneous-slot behavior and the golden-test accessors, shrinking from 791
to 653 lines with no behavior change.
