+++ buildsystem-core/src/main/java/de/eintosti/buildsystem/world/menu/EditMenu.java
@@ -17,8 +17,6 @@
  */
 package de.eintosti.buildsystem.world.menu;
 
-import static java.util.Map.entry;
-
 import com.cryptomorin.xseries.XEntityType;
 import com.cryptomorin.xseries.XMaterial;
 import com.cryptomorin.xseries.XSound;
@@ -41,7 +39,6 @@ import de.eintosti.buildsystem.util.color.ColorAPI;
 import java.util.*;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.function.BiConsumer;
-import java.util.function.Predicate;
 import org.bukkit.*;
 import org.bukkit.entity.Player;
 import org.bukkit.event.inventory.InventoryClickEvent;
@@ -54,7 +51,6 @@ import org.jspecify.annotations.Nullable;
 public class EditMenu extends ButtonMenu<EditMenu.EditButton> {
 
     private static final int SLOT_WORLD_INFO = 3;
-    private static final int SLOT_PIN = 5;
     private static final int SLOT_TIME = 23;
     private static final int SLOT_BUTCHER = 29;
     private static final int SLOT_BUILDERS = 30;
@@ -81,103 +77,6 @@ public class EditMenu extends ButtonMenu<EditMenu.EditButton> {
             XEntityType.SPAWNER_MINECART,
             XEntityType.TNT_MINECART,
             XEntityType.PLAYER);
-
-    /**
-     * Slots whose only action is "check permission, flip a boolean world setting, re-open". They are rendered and
-     * handled uniformly; heterogeneous slots (sub-menus, time, butcher, difficulty) are handled individually.
-     */
-    private static final Map<Integer, Toggle> TOGGLES = Map.ofEntries(
-            entry(
-                    SLOT_PIN,
-                    new Toggle(
-                            XMaterial.ITEM_FRAME,
-                            XMaterial.GLOW_ITEM_FRAME,
-                            "buildsystem.edit.pin",
-                            "worldeditor_pin_item",
-                            "worldeditor_pin_lore",
-                            WorldData::isPinned,
-                            WorldData::setPinned)),
-            entry(
-                    20,
-                    new Toggle(
-                            XMaterial.OAK_PLANKS,
-                            "buildsystem.edit.breaking",
-                            "worldeditor_blockbreaking_item",
-                            "worldeditor_blockbreaking_lore",
-                            WorldData::isBlockBreaking,
-                            WorldData::setBlockBreaking)),
-            entry(
-                    21,
-                    new Toggle(
-                            XMaterial.POLISHED_ANDESITE,
-                            "buildsystem.edit.placement",
-                            "worldeditor_blockplacement_item",
-                            "worldeditor_blockplacement_lore",
-                            WorldData::isBlockPlacement,
-                            WorldData::setBlockPlacement)),
-            entry(
-                    22,
-                    new Toggle(
-                            XMaterial.SAND,
-                            "buildsystem.edit.physics",
-                            "worldeditor_physics_item",
-                            "worldeditor_physics_lore",
-                            WorldData::isPhysics,
-                            WorldData::setPhysics)),
-            entry(
-                    24,
-                    new Toggle(
-                            XMaterial.TNT,
-                            "buildsystem.edit.explosions",
-                            "worldeditor_explosions_item",
-                            "worldeditor_explosions_lore",
-                            WorldData::isExplosions,
-                            WorldData::setExplosions)),
-            entry(
-                    31,
-                    new Toggle(
-                            XMaterial.ARMOR_STAND,
-                            "buildsystem.edit.mobai",
-                            "worldeditor_mobai_item",
-                            "worldeditor_mobai_lore",
-                            WorldData::isMobAi,
-                            WorldData::setMobAi)),
-            entry(
-                    33,
-                    new Toggle(
-                            XMaterial.TRIPWIRE_HOOK,
-                            "buildsystem.edit.interactions",
-                            "worldeditor_blockinteractions_item",
-                            "worldeditor_blockinteractions_lore",
-                            WorldData::isBlockInteractions,
-                            WorldData::setBlockInteractions)));
-
-    private record Toggle(
-            XMaterial material,
-            XMaterial enabledMaterial,
-            String permission,
-            String itemKey,
-            String loreKey,
-            Predicate<WorldData> getter,
-            BiConsumer<WorldData, Boolean> setter) {
-
-        /**
-         * Creates a toggle whose icon is the same whether the setting is enabled or not.
-         */
-        Toggle(
-                XMaterial material,
-                String permission,
-                String itemKey,
-                String loreKey,
-                Predicate<WorldData> getter,
-                BiConsumer<WorldData, Boolean> setter) {
-            this(material, material, permission, itemKey, loreKey, getter, setter);
-        }
-
-        XMaterial iconFor(boolean enabled) {
-            return enabled ? enabledMaterial : material;
-        }
-    }
 
     /**
      * Classifies what a button does after a click, so the per-slot contract can be asserted as data. The actual action
@@ -293,12 +192,13 @@ public class EditMenu extends ButtonMenu<EditMenu.EditButton> {
                         .onClick(this::onWorldInfoClick)
                         .build());
 
-        TOGGLES.forEach((slot, toggle) -> register(
+        EditMenuToggles.TOGGLES.forEach((slot, toggle) -> register(
                 slot,
                 EditButton.builder()
                         .permission(toggle.permission())
                         .outcome(ClickOutcome.REOPEN)
-                        .render((player, inventory) -> renderToggle(player, inventory, slot, toggle))
+                        .render((player, inventory) ->
+                                toggle.render(menuItems, buildWorld.getData(), player, inventory, slot))
                         .onClick((player, event) -> {
                             if (requirePermission(player, toggle.permission())) {
                                 WorldData worldData = buildWorld.getData();
@@ -429,13 +329,6 @@ public class EditMenu extends ButtonMenu<EditMenu.EditButton> {
     protected void populate(Player player) {
         menuItems.fillAll(player, getInventory());
         renderButtons(player);
-    }
-
-    private void renderToggle(Player player, Inventory inventory, int slot, Toggle toggle) {
-        WorldData worldData = buildWorld.getData();
-        boolean enabled = toggle.getter().test(worldData);
-        menuItems.addToggleItem(
-                player, inventory, slot, toggle.iconFor(enabled), enabled, toggle.itemKey(), toggle.loreKey());
     }
 
     private void renderWorldInfo(Player player, Inventory inventory) {
@@ -756,36 +649,5 @@ public class EditMenu extends ButtonMenu<EditMenu.EditButton> {
 
         player.closeInventory();
         messages.sendMessage(player, "worldeditor_butcher_removed", Map.entry("%amount%", entitiesRemoved.get()));
-    }
-
-    /**
-     * Which third of the Minecraft day the world clock currently sits in. Drives the editor's time button.
-     */
-    public enum TimeOfDay {
-        SUNRISE,
-        NOON,
-        NIGHT;
-
-        /**
-         * Minecraft tick at which night begins (the day is 24000 ticks).
-         */
-        static final int NIGHT_START_TICKS = 13000;
-
-        /**
-         * Buckets a raw world tick into a {@link TimeOfDay}.
-         *
-         * @param worldTicks The world time in ticks (0–24000)
-         * @param noonStart The configured tick at which noon begins
-         * @return The matching time-of-day bucket
-         */
-        static TimeOfDay fromTicks(int worldTicks, int noonStart) {
-            if (worldTicks >= 0 && worldTicks < noonStart) {
-                return SUNRISE;
-            } else if (worldTicks >= noonStart && worldTicks < NIGHT_START_TICKS) {
-                return NOON;
-            } else {
-                return NIGHT;
-            }
-        }
     }
 }
